renamed temporary poll results file
