+++ myapp/src/components/NewPolls/PollResults.js
