+++ myapp/src/components/homePagePolls.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import useSWR from "swr";
-import PollResults from "./NewPolls/tempnewpollcopyforhomepage.js";
+import PollResults from "./NewPolls/PollResults.js";
 <script src="https://www.w3schools.com/lib/w3.js"></script>;
 
 const initialList = [];
